Complete 4.21

+++ controllers/blogs.js
@@ -44,7 +44,26 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = request.token
+    ? jwt.verify(request.token, process.env.SECRET)
+    : null;
+  if (!(decodedToken && decodedToken.id)) {
+    return response.status(401).json({ error: 'Token missing or invalid' });
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id);
+  if (blogToDelete.user.toString() !== decodedToken.id) {
+    return response.status(401).json({
+      error: 'Blogs can only be deleted by their creator'
+    });
+  }
+
+  const user = await User.findById(decodedToken.id);
   await Blog.findByIdAndDelete(request.params.id);
+  user.blogs = user.blogs.filter(blog => {
+    return blog.toString() !== blogToDelete._id.toString();
+  });
+  await user.save();
   response.status(204).end();
 });
